ops/backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
dj-rest-auth==1.1.1
Django==3.1.1
django-auth-ldap==2.2.0
django-filter==2.3.0
djangorestframework==3.12.1
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
paramiko==2.7.2
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
python-ldap==3.3.1
pytz==2020.1
requests==2.24.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
zipp==3.1.0
